pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "esxi-mcp-server"
version = "1.0.0"
description = "A Model Context Protocol (MCP) server for managing VMware ESXi/vSphere environments"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "davidroman1914", email = "davidroman1914@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Distributed Computing",
    "Topic :: System :: Clustering",
    "Topic :: System :: Hardware",
]
requires-python = ">=3.11"
dependencies = [
    # MCP Server dependencies
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0",
    
    # VMware SDK - Official VMware vSphere Automation SDK
    "lxml>=4.3.0",
    "six>=1.12",
    "requests>=2.27.1",
    "pyOpenSSL>=22.0.0",
    "cryptography>=37.0.2",
    "setuptools>=62.3.2",
    "pyVmomi==9.0.0.0",
    "vmware-vapi-common-client==2.61.2",
    "vmware-vapi-runtime==2.61.2",
    "vmware-vcenter==9.0.0.0",
    
    # Additional dependencies
    "urllib3>=1.26.0",
    "structlog>=23.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
esxi-mcp-server = "src.__main__:main"

[project.urls]
Homepage = "https://github.com/davidroman1914/vmware-mcp-server"
Repository = "https://github.com/davidroman1914/vmware-mcp-server"
Issues = "https://github.com/davidroman1914/vmware-mcp-server/issues"
Documentation = "https://github.com/davidroman1914/vmware-mcp-server#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pyVim.*",
    "pyVmomi.*",
    "mcp.*",
    "vmware.*",
    "com.vmware.*",
]
ignore_missing_imports = true 